Added server side files and stuff, everything with chat works correctly now
Previously there were problems with messages being sent to the wrong channels which was because I wasn't updating what channel the user was on in the server side.
There is a small problem with using 2 accounts on the same IP but that's inconsequential for now and I'll go over it at a later point maybe, and there's the problem that XOR (and possibly just saving the data) breaks some characters which I'll have to do something about later, but this does the job for now.